LayoutTests/platform/ios-sim/storage/indexeddb/noblobs-expected.txt

CONSOLE MESSAGE: line 19: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(['resources/test-data.html', 'resources/test-data.txt'])')
Confirm Blob/File/FileList limitations of WebKit's IndexedDB implementation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Unexpected error: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(['resources/test-data.html', 'resources/test-data.txt'])')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
